Replace Audio package with single header/cpp in core
