+++ Sprocket/Core/sound.cpp
@@ -4,7 +4,52 @@
 #include <SFML/Audio.hpp>
 
 namespace spkt {
-namespace Audio {
+
+void set_listener(const glm::vec3& postiion, const glm::vec3& direction)
+{
+    sf::Listener::setPosition(postiion.x, postiion.y, postiion.z);
+    sf::Listener::setDirection(direction.x, direction.y, direction.z);
+}
+
+float get_master_volume()
+{
+    return sf::Listener::getGlobalVolume();
+}
+
+void set_master_volume(float volume)
+{
+    sf::Listener::setGlobalVolume(volume);
+}
+
+struct MusicImpl
+{
+    sf::Music music;
+};
+
+music::music()
+    : d_impl(std::make_unique<MusicImpl>())
+{
+}
+
+bool music::load(const std::string& filename)
+{
+    return d_impl->music.openFromFile(filename);
+}
+
+void music::play() const
+{
+    d_impl->music.play();
+}
+
+void music::pause() const
+{
+    d_impl->music.pause();
+}
+
+void music::stop() const
+{
+    d_impl->music.stop();
+}
 
 struct SoundImpl
 {
@@ -62,5 +107,4 @@ void Source::Stop() const
     d_impl->sound.stop();
 }
 
-}
 }
